refactor: review lib files
