+++ libs/conversation-messages/src/components/MessageActions/tests/MessageActions.spec.tsx
@@ -2,7 +2,7 @@ import { MessageRating, MessageRole } from '@epam/ai-dial-chat-shared';
 import { render, screen } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { describe, expect, it, vi } from 'vitest';
-import { MessageActions } from './MessageActions';
+import { MessageActions } from '../MessageActions';
 
 describe('MessageActions', () => {
   describe('role=User (default)', () => {
