+++ libs/conversation-messages/src/index.ts
@@ -2,7 +2,7 @@ export { MessageBubble } from './components/MessageBubble/MessageBubble';
 export { UserMessageBubble } from './components/MessageBubble/UserMessageBubble';
 export { AssistantMessageBubble } from './components/MessageBubble/AssistantMessageBubble';
 export { StatusMessageBubble } from './components/MessageBubble/StatusMessageBubble';
-export { MessageActions } from './components/Message/MessageActions';
+export { MessageActions } from './components/MessageActions/MessageActions';
 export { MessageSource } from './components/MessageSource/MessageSource';
 export { BubblePosition } from './types/bubble-position';
 export type {
@@ -12,16 +12,16 @@ export type {
   MessageBubbleStyles,
   MessageBubbleColors,
   MessageBubbleTypography,
-} from './models/MessageBubble';
+} from './models/message-bubble';
 export type { StatusMessageBubbleProps } from './components/MessageBubble/StatusMessageBubble';
 export type {
   MessageActionsProps,
   MessageActionTooltips,
   MessageActionAriaLabels,
-} from './models/MessageActions';
+} from './models/message-actions';
 export type {
   MessageSourceProps,
   MessageSourceStyles,
   MessageSourceColors,
   MessageSourceTypography,
-} from './models/MessageSource';
+} from './models/message-source';
