+++ libs/conversation-messages/src/models/message-source.ts
@@ -30,14 +30,14 @@ export interface MessageSourceStyles {
   colors?: MessageSourceColors;
   /** Typography overrides applied via CSS custom properties. */
   typography?: MessageSourceTypography;
+  /** Extra class name(s) merged onto the button element. */
+  className?: string;
 }
 
 /** Props accepted by the `MessageSource` component. */
 export interface MessageSourceProps {
   /** Text label displayed inside the button. */
   label: string;
-  /** Extra class name(s) merged onto the button element. */
-  className?: string;
   /** Called when the user clicks the source button. */
   onClick?: () => void;
   /** Color and typography overrides applied as CSS custom properties. */
